+++ js/util.js
@@ -1,0 +1,1 @@
+export const isValidStringLength = (string, maxLength) => string.length <= maxLength;
